+++ src/main/java/com/coderscampus/flightTrack/repository/RefreshTokenRepository.java
@@ -3,6 +3,6 @@ package com.coderscampus.flightTrack.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.coderscampus.flightTrack.domain.RefreshToken;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long>{
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer>{
 
 }
